Damage states working
